+++ src/index.ts
@@ -6,5 +6,6 @@ export type {ToastContextType,ToastItem} from '../src/toast/types/types'
 export { default as ToastProvider } from './toast/toast-provider/ToastProvider';
 export { useToast } from './toast/hook/useToast'
 export { default as toastware } from './toast/toaster';
+export { default as toaster} from './toast/toaster'