+++ src/toast/toaster.tsx
@@ -1,31 +1,38 @@
 import { getToastAPI } from "./core/globalStore";
+import promise from "./prmosie-toast/promiseToast";
 
 const toaster = {
   addToast: (...args: Parameters<NonNullable<ReturnType<typeof getToastAPI>["addToast"]>>) => {
-    getToastAPI().addToast?.(...args);
+   return getToastAPI().addToast?.(...args);
   },
   notify:(...args: Parameters<NonNullable<ReturnType<typeof getToastAPI>["addToast"]>>) => {
-    getToastAPI().addToast?.(...args);
+    return getToastAPI().addToast?.(...args);
   
   },
-  removeToast: (id: string) => {
+  removeToastById: (id: string) => {
     getToastAPI().removeToast?.(id);
   },
-  clearToasts: () => {
-    getToastAPI().clearToasts?.();
-  },
+ 
   success: (msg: string, pos?: string) => {
-    getToastAPI().addToast?.(msg, "success", pos);
+   return getToastAPI().addToast?.(msg, "success", pos);
+  },
+  warning: (msg: string, pos?: string) => {
+   return getToastAPI().addToast?.(msg, "warning", pos);
   },
   error: (msg: string, pos?: string) => {
-    getToastAPI().addToast?.(msg, "error", pos);
+   return getToastAPI().addToast?.(msg, "error", pos);
   },
   info: (msg: string, pos?: string) => {
-    getToastAPI().addToast?.(msg, "info", pos);
+   return getToastAPI().addToast?.(msg, "info", pos);
+  },
+   clearToasts: () => {
+    getToastAPI().clearToasts?.();
   },
   clear : ()=>{
     getToastAPI().clear?.();
-  }
+  },
+
+  promise :promise
 };
 
 export default toaster;